requirements.txt
mysql-connector-python
pymysql
sqlalchemy
python-dotenv
